Updating code
Configuring the back-end functionality behind login page, register page and the logout button. Also restricting users from deleting rooms they did not create.

+++ django1/studybud/base/views.py
@@ -1,4 +1,10 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.db.models import Q
+from django.contrib.auth import authenticate, login,logout
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -8,12 +14,45 @@ from .forms import RoomForm
 #    {'id':3, 'name':'Frontend Developers'}
 #]
 
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == "POST":
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request,"User does not exist, Please sign up")
+            
+        user = authenticate(request, username=username, password=password)
+        
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'Username or Password is incorrect')
+            
+    context = {'page':page}
+    return render(request, 'base/login_register.html', context)
+
+def LogoutUser(request):
+    logout(request)
+    return redirect('login')
+
+def registerUser(request):
+    page = 'register'
+    return render(request,'base/login_register.html')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
-    rooms = Room.objects.filter(topic__name__icontains=q)
+    rooms = Room.objects.filter(Q(topic__name__icontains=q)|Q(name__icontains=q)|Q(description__icontains=q))
     topics = Topic.objects.all()
-    context = {'rooms':rooms, 'topics':topics}
+    room_count = rooms.count()
+    context = {'rooms':rooms, 'topics':topics, 'room_count':room_count}
     return render(request, 'base/home.html', context) 
 
 def room(request,pk):
@@ -21,6 +60,7 @@ def room(request,pk):
     context = {'room':room}
     return render(request,'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -31,10 +71,13 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('you are not allowed here!!')
     
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -45,8 +88,13 @@ def updateRoom(request,pk):
     context={'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    
+     
+    if request.user != room.host:
+        return HttpResponse('you are not allowed here!!')
     if request.method == "POST":
         room.delete()
         return redirect('home')
